#2 StringSchemaPatternBuilder | Complete flow for processing StringSchemas, and test class with StringSchemaPatternBuilder
StringSchemaPatternBuilder now processes example,
enum, default values, and has matching for regexp
pattern as well as strings with specific lengths.

Added pattern builders for EmailSchemas and UUIDSchemas.

The addition of the test class required an additional
dependency in the build.gradle file, which required
changing the port for the WireMockService that spins up
upon starting the Spring Boot application.

Pattern builders were renamed from factories.

The SchemaToPatternBuilderDispatcher was changed to have
a list helping the injection of pattern builder classes
instead of a HashMap which may be faster but is much
harder to read. Besides, there are only 17 inheritors
of the Schema class.